修复管理员ID 角色ID为空

+++ src/admin/model/SysGroup.php
@@ -14,14 +14,14 @@ class SysGroup extends Model
     public static function onAfterWrite($model)
     {
         Cache::delete('SysGroups');
-        if (!empty($model['admin_ids'])) {
+        if (isset($model['admin_ids']) && !empty($model['admin_ids'])) {
             if (is_string($model['admin_ids'])) {
                 $model['admin_ids'] = explode(',', $model['admin_ids']);
             }
             $model->adminUsers()->detach($model->adminUsers()->column('id'));
             $model->adminUsers()->attach($model['admin_ids'], ['administrator' => 1]);
         }
-        if (!empty($model['role_ids'])) {
+        if (isset($model['role_ids']) && !empty($model['role_ids'])) {
             if (is_string($model['role_ids'])) {
                 $model['role_ids'] = explode(',', $model['role_ids']);
             }
